% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_metadata.R
\name{clean_metadata}
\alias{clean_metadata}
\title{Funkcja czyszczaca metadane}
\usage{
clean_metadata(adres, rzad = "synop", interwal = "terminowe")
}
\arguments{
\item{adres}{adres URL do pliku z metadanymi}
}
\description{
Tylko do wewnetrznego uzytku...
Na potrzeby funkcji 'metadane()'
}
\examples{
# przykladowe pobranie:
\dontrun{
  clean_metadata(adres = "https://dane.imgw.pl/data/dane_pomiarowo_obserwacyjne/dane_meteorologiczne/dobowe/synop/s_d_format.txt")
}

}
