% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_metadata_hydro.R
\name{clean_metadata_hydro}
\alias{clean_metadata_hydro}
\title{Funkcja czyszczaca hydro metadane}
\usage{
clean_metadata_hydro(adres, interwal = "miesieczne")
}
\arguments{
\item{adres}{adres URL do pliku z metadanymi}

\item{interwal}{interwal czasowy}
}
\description{
Tylko do wewnetrznego uzytku...
Na potrzeby funkcji "hydro()" ?
}
\examples{
\dontrun{

}
}
