% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_metadata_meteo.R
\name{clean_metadata_meteo}
\alias{clean_metadata_meteo}
\title{Funkcja czyszczaca metadane}
\usage{
clean_metadata_meteo(address, rank = "synop", interval = "terminowe")
}
\arguments{
\item{address}{address URL to file with metadata}

\item{rank}{rank}

\item{interval}{interval czasowy}
}
\description{
Tylko do wewnetrznego uzytku...
Na potrzeby funkcji 'meteo_metadane()'
}
\examples{
\dontrun{
  clean_metadata_meteo(address = "https://dane.imgw.pl/data/dane_pomiarowo_obserwacyjne/dane_meteorologiczne/dobowe/synop/s_d_format.txt", rank = "synop", interval = "terminowe")
}

}
