% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_metadata_miesieczne.R
\name{clean_metadata_miesieczne}
\alias{clean_metadata_miesieczne}
\title{Funkcja czyszczaca metadane dla plikow miesiecznych}
\usage{
clean_metadata_miesieczne(adres)
}
\arguments{
\item{adres}{adres URL do pliku z metadanymi}
}
\description{
Tylko do wewnetrznego uzytku...
Na potrzeby funkcji 'metadane()'
}
