% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro.R
\name{hydro}
\alias{hydro}
\title{Hydrological data}
\usage{
hydro(interval, year, value = NULL)
}
\arguments{
\item{interval}{temporal resolution of the data ("daily" , "monthly", or "semiannual_and_annual")}

\item{year}{vector of years (e.g., 1966:2000)}

\item{value}{type of data (can be: state - "H" (default), flow - "Q", or temperature - "T")}
}
\description{
Downloading hourly, daily, and monthly hydrological data from the SYNOP / CLIMATE / PRECIP stations available in the danepubliczne.imgw.pl collection
}
\examples{
\dontrun{
  x <- hydro("monthly", year = 1999)
  head(x)
}
}
