% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_annual.R
\name{hydro_annual}
\alias{hydro_annual}
\title{Semi-annual and annual hydrological data}
\usage{
hydro_annual(year = 1966:2000, coords = FALSE, value = "H")
}
\arguments{
\item{year}{vector of years (e.g. 1966:2000)}

\item{coords}{add coordinates for the station (logical value TRUE or FALSE)}

\item{value}{type of data ( can be: "State=H (default)" , "Flow=Q" , "Temperature=T")}
}
\description{
Downloading hydrological data for the semi-annual and annual period
available in the collection danepubliczne.imgw.pl
}
\examples{
\dontrun{
  yearly <- hydro_annual(year = 2000, value="H")
  head(yearly)
}

}
