% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_clean_meta.R
\name{hydro_clean_metadata}
\alias{hydro_clean_metadata}
\title{Funkcja czyszczaca hydro metadane}
\usage{
hydro_clean_metadata(adres, interwal = "miesieczne")
}
\arguments{
\item{adres}{adres URL do pliku z metadanymi}

\item{interwal}{interwal czasowy}
}
\description{
Tylko do wewnetrznego uzytku...
Na potrzeby funkcji "hydro()" ?
}
\examples{
\dontrun{

}
}
