% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_daily.R
\name{hydro_daily}
\alias{hydro_daily}
\title{Daily hydrological data}
\usage{
hydro_daily(year = 1966:2000, coords = FALSE)
}
\arguments{
\item{year}{vector of years (e.g., 1966:2000)}

\item{coords}{add coordinates of the station (logical value TRUE or FALSE)}
}
\description{
Downloading daily hydrological data from the danepubliczne.imgw.pl collection
}
\examples{
\dontrun{
  daily <- hydro_daily(year = 2000)
  head(mies)
}

}
