% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_dobowe.R
\name{hydro_dobowe}
\alias{hydro_dobowe}
\title{Daily hydrological data}
\usage{
hydro_dobowe(rok = 1966:2000, coords = FALSE)
}
\arguments{
\item{rok}{wektor dla wybranych lat (np. 1966:2000)}

\item{coords}{czy dodac wspolrzedne (WGS84) i wysokosc stacji? (domyslnie coords = FALSE)}
}
\description{
Pobranie danych hydrologicznych dla dobowego okresu
udostepnionego w zbiorze danepubliczne.imgw.pl
}
\examples{
\dontrun{
  dobowe <- hydro_dobowe(rok = 2000)
  head(mies)
}

}
