% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_metadane.R
\name{hydro_metadane}
\alias{hydro_metadane}
\title{Pobranie opisu (metadanych) do danych hydrologicznych  udostępnionych w repozytorium danepubliczne.imgw.pl
Domyślnie funkcja zwraca listę bądź ramkę danych dla wybranego podzbioru (w zależności od radzaj)}
\usage{
hydro_metadane(interwal)
}
\arguments{
\item{interwal}{\itemize{
\item Argument przyjmujący wartość: 'dobowe' , miesieczne' lub 'polroczne_i_roczne'.  Oznacza dla jakiej rozdzielczosci czasowej maja zostac przygotowane metadane dla srodowiska R
}}
}
\description{
Pobranie opisu (metadanych) do danych hydrologicznych  udostępnionych w repozytorium danepubliczne.imgw.pl
Domyślnie funkcja zwraca listę bądź ramkę danych dla wybranego podzbioru (w zależności od radzaj)
}
\examples{
# przykladowe uzycie:
\dontrun{
meta <- metadane(interwal = "dobowe")
meta <- metadane(interwal = "miesieczne")
meta <- metadane(interwal = "polroczne_i_roczne")
}
}
