% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_metadane.R
\name{hydro_metadata}
\alias{hydro_metadata}
\title{Hydrological metadata}
\usage{
hydro_metadata(interval)
}
\arguments{
\item{interval}{\itemize{
\item Argument accepts the values: 'dobowe' , miesieczne' or 'polroczne_i_roczne'.  Indicates for which time resolution metadata for environment R should be prepared
}}
}
\description{
Downloading the description (metadata) to hydrological data available in the danepubliczne.imgw.pl repository
By default, the function returns a list or data frame for a selected subset (depending on how to cope)
}
\examples{
\dontrun{
  meta <- hydro_metadata(interval = "dobowe")
  meta <- hydro_metadata(interval = "miesieczne")
  meta <- hydro_metadata(interval = "polroczne_i_roczne")
}
}
