% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_miesieczne.R
\name{hydro_miesieczne}
\alias{hydro_miesieczne}
\title{Monthly hydrological data}
\usage{
hydro_miesieczne(rok = 1966:2000, coords = FALSE)
}
\arguments{
\item{rok}{wektor dla wybranych lat (np. 1966:2000)}

\item{coords}{czy dodac wspolrzedne (WGS84) i wysokosc stacji? (domyslnie coords = FALSE)}
}
\description{
Pobranie danych hydrologicznych dla miesiecznego okresu, udostepnionych w zbiorze danepubliczne.imgw.pl
}
\examples{
\dontrun{
  mies <- hydro_miesieczne(rok = 2000)
  head(mies)
}

}
