% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_monthly.R
\name{hydro_monthly}
\alias{hydro_monthly}
\title{Monthly hydrological data}
\usage{
hydro_monthly(year, station = NULL)
}
\arguments{
\item{year}{vector of years (e.g., 1966:2000)}

\item{station}{name or ID of hydrological station(s).
It accepts names (characters in CAPITAL LETTERS) or stations' IDs (numeric)}
}
\description{
Downloading monthly hydrological data from the danepubliczne.imgw.pl collection
}
\examples{
\dontrun{
  monthly <- hydro_monthly(year = 2000)
  head(monthly)
}

}
