% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_roczne.R
\name{hydro_roczne}
\alias{hydro_roczne}
\title{Pobranie danych hydrologicznych dla rocznego okresu
udostepnionego w zbiorze danepubliczne.imgw.pl}
\usage{
hydro_roczne(rok = 1966:2000, value = "H", coords = FALSE)
}
\arguments{
\item{rok}{wektor dla wybranych lat hydrologicznych (np. 1966:2000)}

\item{value}{rodzaj danych (do wyboru: "Stan=H (domyslnie)" , "Przepływ=Q" , "Temperatura=T")}

\item{coords}{czy dodac wspolrzedne (WGS84) i wysokosc stacji? (domyslnie coords = FALSE)}
}
\description{
Pobranie danych hydrologicznych dla rocznego okresu
udostepnionego w zbiorze danepubliczne.imgw.pl
}
\examples{
\dontrun{
roczne <- hydro_roczne(rok = 2000,value="H")
  head(mies)
}

}
