% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_yearly.R
\name{hydro_yearly}
\alias{hydro_yearly}
\title{Semiannual and annual hydrological data}
\usage{
hydro_yearly(year = 1966:2000, coords = FALSE, value = "H")
}
\arguments{
\item{year}{vector of years (e.g. 1966:2000)}

\item{coords}{add coordinates for the station (logical value TRUE or FALSE)}

\item{value}{type of data ( can be: "State=H (default)" , "Flow=Q" , "Temperature=T")}
}
\description{
Downloading hydrological data for the annual period
available in the collection danepubliczne.imgw.pl
}
\examples{
\dontrun{
  yearly <- hydro_yearly(year = 2000,value="H")
  head(mies)
}

}
