% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imgw-package.R
\docType{package}
\name{imgw-package}
\alias{imgw}
\alias{imgw-package}
\title{imgw: imgw meteorological data downloader}
\description{
A set of functions for downloading Polish meteorological data from the IMGW-PIB.
}
\author{
\strong{Maintainer}: Bartosz Czernecki \email{nwp@amu.edu.pl}

}
\keyword{internal}
