% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadane.R
\name{metadane}
\alias{metadane}
\title{Funkcja metadane}
\usage{
metadane(okres, rodzaj = "synop")
}
\arguments{
\item{okres}{\itemize{
\item argumenty przyjmujący wartość: 'dobowe', 'miesieczne' badz 'terminowe'.  Oznacza dla jakiej rozdzielczosci czasowej maja zostac przygotowane metadane dla srodowiska R
}}

\item{rodzaj}{\itemize{
\item argumenty przyjmujący wartość: 'synop' (domyślnie) lub 'klimat'.  Oznacza dla jakiego rodzaju/rzędu staji mają zostać przygotowane metadane.
}}
}
\description{
Funkcja metadane
}
\examples{
# przykladowe uzycie:
\dontrun{
meta <- metadane(okres="dobowe")
meta <- metadane(okres="terminowe")
meta <- metadane(okres="miesieczne")
}
}
