% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo_hourly.R
\name{meteo_hourly}
\alias{meteo_hourly}
\title{Hourly meteorological data}
\usage{
meteo_hourly(rank = "synop", year = 1966:2018, status = FALSE,
  coords = FALSE)
}
\arguments{
\item{rank}{rank of station ("synop" , "klimat" , "opad")}

\item{year}{vector of years (np. 1966:2000)}

\item{status}{leave the columns with measurement or observation statuses (default status = FALSE - i.e. the status columns are deleted)}

\item{coords}{add coordinates for the station (logical value TRUE or FALSE)}
}
\description{
Downloading term (meteorological) data from the SYNOP / KLIMAT station available in the danepubliczne.imgw.pl collection
}
\examples{
\dontrun{
  hourly <- meteo_hourly(rank = "klimat", year = 1970:2000)
  head(terminowe)
}

}
