% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo_monthly.R
\name{meteo_monthly}
\alias{meteo_monthly}
\title{Monthly meteorological data}
\usage{
meteo_monthly(rank = "synop", year = 1966:2018, status = FALSE,
  coords = FALSE)
}
\arguments{
\item{rank}{rank of station ("synop" , "klimat" , "opad")}

\item{year}{vector of years (np. 1966:2000)}

\item{status}{leave the columns with measurement or observation statuses (default status = FALSE - i.e. the status columns are deleted)}

\item{coords}{add coordinates for the station (logical value TRUE or FALSE)}
}
\description{
Downloading monthly (meteorological) data from SYNOP / KLIMAT / OPAD stations available in the danepubliczne.imgw.pl collection
}
\examples{
\dontrun{
  monthly <- meteo_monthly(rank = "klimat")
  head(monthly)
}

}
