% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo_shortening.R
\name{meteo_shortening}
\alias{meteo_shortening}
\title{Shortening column names for meteorological variables}
\usage{
meteo_shortening(data, format = "short", duplicates = TRUE)
}
\arguments{
\item{data}{downloaded dataset with original column names}

\item{format}{three types of column names possible: "short" - default, values with shorten names, "polish" - (original names as downloaded), "full" - full English description}

\item{duplicates}{whether to remove duplicated column names (default TRUE - i.e., columns with duplicated names are deleted)}
}
\description{
Shortening column names of meteorological parameters to improve the readability of downloaded dataset and removing duplicated column names
}
\examples{
\dontrun{
  colnames(data)
  abbr <- meteo_shortening(data = data, format = "short", duplicates = TRUE)
  head(abbr)
}

}
