% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo_sounding.R
\name{meteo_sounding}
\alias{meteo_sounding}
\title{Sounding data downloader}
\usage{
meteo_sounding(wmo_id = 12120, yy = 2019, mm = 4, dd = 9, hh = 0)
}
\arguments{
\item{wmo_id}{międzynarodowy kod stacji wg WMO (World Meteorological Organization); Dla polskich stacji: Łeba - 12330, Legionowo - 12374, Wrocław- 12425}

\item{yy}{rok}

\item{mm}{czy usunac kolumny ze statusami pomiarow lub obserwacji (domyslnie status = FALSE - tj. kolumny ze statusami sa usuwane)}

\item{dd}{czy dodac wspolrzedne (WGS84) i wysokosc stacji? (domyslnie coords = FALSE)}

\item{hh}{na większość stacji do wyboru godziny 00 i 12; sporadczynie pomiary wykonywane co 6h}
}
\description{
Pobranie danych z pionowych sondowań atmosfery; Dane oryginalnie dostępne w serwisie: http://weather.uwyo.edu/upperair/sounding.html
}
\examples{
\dontrun{
  sounding <- meteo_sounding()
  head(sounding)
}

}
