% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo_terminowe.R
\name{meteo_terminowe}
\alias{meteo_terminowe}
\title{Hourly meteorological data}
\usage{
meteo_terminowe(rank = "synop", rok = 1966:2018, status = FALSE,
  coords = FALSE)
}
\arguments{
\item{rank}{rzad stacji (do wyboru: "synop", "klimat")}

\item{rok}{wektor dla wybranych lat (np. 1966:2000)}

\item{status}{czy pozostawic kolumny ze statusami pomiarow lub obserwacji (domyslnie status = FALSE - tj. kolumny ze statusami sa usuwane )}

\item{coords}{czy dodac koordynaty dla stacji (wartosc logiczna TRUE lub FALSE)}
}
\description{
Pobranie danych terminowych  (meteorologicznych) ze stacji SYNOP/KLIMAT udostepnionych w zbiorze danepubliczne.imgw.pl
}
\examples{
\dontrun{
  terminowe <- meteo_terminowe(rzad = "klimat", lata = 1970:2000)
  head(terminowe)
}

}
