% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor_parse.R
\name{monitor_imgw}
\alias{monitor_imgw}
\title{Funkcja pozwalająca na pobranie danych umieszczonych w monitorze IMGW-PIB}
\usage{
monitor_imgw(adres_monitor)
}
\arguments{
\item{adres_monitor}{adres URL strony Monitora IMGW-PIB (np. http://monitor.pogodynka.pl/#station/meteo/352160330)}
}
\description{
Dane meteorologiczne i hydrologiczne zawarte w serwisie monitor IMGW
są dostarczane przez Instytut Meteorologii i Gospodarki Wodnej - Państwowy Instytutu Badawczy
}
\examples{
# przykladowe pobranie:
adres_monitor <- "http://monitor.pogodynka.pl/#station/meteo/352160330"
monitor_imgw(adres_monitor)
}

