% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogimet_hourly.R
\name{ogimet_hours}
\alias{ogimet_hours}
\title{Hourly meteorological data form ogimet}
\usage{
ogimet_hours(date = c("2018-05-01", "2018-07-01"), coords = FALSE,
  station = c(12326, 12195), col_names = "short", ...)
}
\arguments{
\item{date}{start and finish of date (e.g., date=c("2018-05-01","2018-07-01") )}

\item{coords}{add coordinates of the station (logical value TRUE or FALSE)}

\item{station}{ID of meteorological station(s).
It accepts stations' IDs (numeric)}

\item{col_names}{three types of column names possible: "short" - default, values with shorten names, "full" - full English description, "polish" - original names in the dataset}

\item{...}{other parameters that may be passed to the 'shortening' function that shortens column names}
}
\description{
Downloading houly (meteorological) data from the stations available in the https://www.ogimet.com/ collection
}
\examples{
\donttest{
  daily_ogi <- ogimet_hours()
  head(daily_ogi)
}

}
