% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pobierz_dobowe.R
\name{pobierz_dobowe}
\alias{pobierz_dobowe}
\title{Funkcja pobierająca dane miesięczne}
\usage{
pobierz_dobowe()
}
\description{
Funkcja pobiera wszystkie dostępne dane miesięczne ze stacji SYNOP dostępne w zbiorze danepubliczne.imgw.pl
Domyślnie dane zapisywane są w aktualnie wskazanym katalogu roboczym.
Po pobraniu danych nalezy jest nastepnie rozpakowac i wczytac za pomoca funkcji 'wczytaj_dobowe()'
}
\examples{
pobierz_dobowe() # download all files into current working directory
# next steps should involve (1) unpacking and (2) reading into R environment:
# wczytaj_dobowe()
}
