% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pobierz_dobowe2.R
\name{pobierz_dobowe2}
\alias{pobierz_dobowe2}
\title{Funkcja pobierająca dane miesięczne ale tylko dla wskazanego folderu/roku}
\usage{
pobierz_dobowe2(rok = "1996_2000")
}
\description{
Funkcja pobiera wszystkie dostępne dane miesięczne ze stacji SYNOP dostępne w zbiorze danepubliczne.imgw.pl
Domyślnie dane zapisywane są w aktualnie wskazanym katalogu roboczym.
Po pobraniu danych nalezy jest nastepnie rozpakowac i wczytac za pomoca funkcji 'wczytaj_dobowe()'
}
