% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pobierz_lapply.R
\name{pobierz_lapply}
\alias{pobierz_lapply}
\title{Pobieranie danych meteorologicznych z serwisu IMGW-PIB (https://dane.imgw.pl)}
\usage{
pobierz_lapply(data_start, data_end, user_pass, stacja, kod, cores)
}
\arguments{
\item{data_start}{Data początkowa od której będą pobierane dane. Obiekt klasy Date}

\item{data_end}{Data końcowa do której będą pobierane. Obiekt klasy Date. Domyślnie funkcja rozszerza zakres pobierania danych o tydzień w przód i w tył}

\item{user_pass}{Login i token wygenerowany przez serwis https://dane.imgw.pl. Parametr w postaci email:token, np. bczernecki@gmail.com:68a60055}

\item{stacja}{Kod stacji wg strony https://dane.imgw.pl}

\item{kod}{Kod pobieranego elementu meteorologicznego/hydrologicznego}

\item{cores}{Liczba wątków do zrównoleglonego pobierania}
}
\description{
Funkcja pozwalająca na zautomatyzowanie procesu pobierania danych meteorologicznych i hydrologicznych z oficjalnego repozytorium Instytutu Meteorologii i Gospodarki Wodnej - Państwowego Instytutu Badawczego - https://dane.imgw.pl.
Jest to zrównoleglona wersja funkcji 'pobierz'. Z uwagi na możliwość otwarcia max. 128 połączeń (textConnections) w R liczba wykonywanych zapytań jest dzielona na proporcjonalnie mniejsze części.
}
\examples{
# wymagane ustalenie ponizszych parametrow pobierania:

user_pass <- "" # pamietaj o wpisaniu swojego loginu:tokena wg schematu zawartego w opisie!
data_start <- "2016-12-26" 
data_end <- "2017-02-20" 
stacja <- "352160330" # przykladowe id dla Poznania; numer mozna pobrac ze strony https://dane.imgw.pl
kod <- "B100B008CD" # srednia dobowa temp. powietrza synop; j.w.
cores <- 4 # liczba watkow do mclapply
pobierz_lapply(data_start,data_end,user_pass,stacja,kod,cores)
}
