% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pobierz_miesieczne.R
\name{pobierz_miesieczne}
\alias{pobierz_miesieczne}
\title{Funkcja pobierająca dane miesięczne}
\usage{
pobierz_miesieczne()
}
\description{
Funkcja pobierająca wszystkie dostępne dane miesięczne ze stacji SYNOP dostępne w zbiorze danepubliczne.imgw.pl
}
\examples{
miesieczne <- pobierz_miesieczne
head(miesieczne)
}
