% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pobierz_terminowe.R
\name{pobierz_terminowe}
\alias{pobierz_terminowe}
\title{Funkcja pobierająca dane terminowe}
\usage{
pobierz_terminowe()
}
\description{
Funkcja pobiera wszystkie dostępne dane terminowe ze stacji SYNOP dostępne w zbiorze danepubliczne.imgw.pl
W obecnej (mocno roboczej postaci) funkcja pobiera wszystkie dane, co wymaga dostepu do duzej ilosc pamieci RAM (~kilkadziesiat GB)
}
\examples{
result <- pobierz_terminowe() # currently requires lot of memory!
}
