% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pobierz_terminowe2.R
\name{pobierz_terminowe2}
\alias{pobierz_terminowe2}
\title{Funkcja pobierająca dane terminowe ale tylko dla wybranego roku / wielolecia}
\usage{
pobierz_terminowe2(rok = 2017)
}
\description{
Funkcja pobiera dane terminowe ze stacji SYNOP dostępne w jednym z katalogow pod adresem https://dane.imgw.pl/data/dane_pomiarowo_obserwacyjne/dane_meteorologiczne/terminowe/synop/
Jako argument nalezy podac POJEDYNCZY rok (>=2001) lub wielolecie (dla danych sprzed 2000)
}
\examples{
result <- pobierz_terminowe2(rok=2005) 
# jesli chcemy pobrac wiecej niz 1 rok warto skorzystac z funkcji 'lapply'
# ponizsza operacja automatycznie pobierze lata 2005-2007:
wynik <- lapply(2005:2007, pobierz_terminowe2)
library(data.table)
wynik2 <- rbindlist(wynik)
}

