% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.fwf2.R
\name{read.fwf2}
\alias{read.fwf2}
\title{For internal use only
Na potrzeby funkcji 'clean_metadane()' (pozbycie sie warningow)}
\usage{
read.fwf2(file, widths, header = FALSE, sep = "\\t", skip = 0L,
  row.names, col.names, n = -1L, buffersize = 2000,
  fileEncoding = "", ...)
}
\description{
For internal use only
Na potrzeby funkcji 'clean_metadane()' (pozbycie sie warningow)
}
