% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortening.R
\name{shortening}
\alias{shortening}
\title{Shortening column names}
\usage{
shortening(data, col_names = "short", duplicates = TRUE)
}
\arguments{
\item{data}{downloaded dataset with original column names}

\item{col_names}{three types of column names possible: "short" - default, values with shorten names, "full" - full English description, "polish" - original names in the dataset}

\item{duplicates}{whether to remove duplicated column names (default TRUE - i.e., columns with duplicated names are deleted)}
}
\description{
Shortening column names to improve the readability of downloaded dataset and removing duplicated column names
}
\examples{
\dontrun{
  colnames(data)
  abbr <- shortening(data = data, colnames = "short", duplicates = TRUE)
  head(abbr)
}

}
