% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacje_meteo.R
\docType{data}
\name{stacje_meteo}
\alias{stacje_meteo}
\title{Location of the weather stations}
\format{The data contains a data.frame with 1998 obs. of 4 variables:
\itemize{
\item{X} {Longitude}
\item{Y} {Latitude}
\item{Kod_stacji} {Station ID}
\item{Wysokosc} {Station elevation in m asl}
}
The object is in the geographic coordinates using WGS84 (EPSG:4326).}
\usage{
stacje_meteo
}
\description{
The object contains weather stations
coordinates, ID numbers, and elevations
}
\examples{
data(stacje_meteo)
head(stacje_meteo)
}
\keyword{datasets}
\keyword{meteo}
