% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wczytaj_miesieczne.R
\name{wczytaj_miesieczne}
\alias{wczytaj_miesieczne}
\title{Funkcja wczytujaca pobrane dane}
\usage{
wczytaj_miesieczne(katalog)
}
\arguments{
\item{katalog}{Wskazany katalog z plikami .zip}
}
\description{
Funkcja wczytuje wszystkie (uprzednio pobrane) pliki z rozszerzeniem '.zip' w katalogu roboczym (lub wskazanym).
Komenda rozpakowywuje wszystkie pliki .zip, w ktorych domyslnie znajduja sie pliki .csv (s_d_t... oraz s_d...),
wczytuje je do oddzielnych list bedacych wynikiem dzialania funkcji
}
\details{
Uwaga! Inne pliki z rozszerzeniem .zip w katalogu roboczym rowniez zostana rozpakowane!
Uwaga2! W przypadku duzej liczby plikow caly proces moze byc czaso- i obliczeniochlonny!
}
\examples{
a <- (dir(pattern=".zip")) # tu mam 1500 plikow zip, z ktorych kazdy zawiera ~2-3 pliki w srodku
b <- as.list(a)    # robie z tego liste
lapply(b, FUN = function(x) unzip(zipfile = x)) # i wszystko rozpakowywuje
# next steps should involve (1) unpacking and (2) reading into R environment:

dane <- wczytaj_miesieczne(katalog = "/home/bartosz/github/imgw/inst/extdata/dobowe/")
}
