% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abbrev.R
\name{abbrev}
\alias{abbrev}
\title{Shortening column names}
\usage{
abbrev(data, duplicates = TRUE, format = "short")
}
\arguments{
\item{data}{downloaded dataset with original column names}

\item{duplicates}{whether to remove duplicated column names (default TRUE - i.e. columns with duplicated names are deleted)}

\item{format}{three types of column names possible: "short" - default, values with shorten names, "polish" - (original names as downloaded), "full" - full English description}
}
\description{
Shortening column names to improve readability of downloaded dataset and removing duplicated column names
}
\examples{
\dontrun{
  abbrev <- abbre(rank = "climate", year = 1969)
  head(monthly)
}

}
