% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_metadata_hydro.R
\name{clean_metadata_hydro}
\alias{clean_metadata_hydro}
\title{Funkcja czyszczaca hydro metadane}
\usage{
clean_metadata_hydro(address, interval = "miesieczne")
}
\arguments{
\item{address}{address URL do pliku z metadanymi}

\item{interval}{interval czasowy}
}
\description{
Tylko do wewnetrznego uzytku...
Na potrzeby funkcji "hydro()" ?
}
\examples{
\dontrun{

}
}
