% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_metadata_meteo.R
\name{clean_metadata_meteo}
\alias{clean_metadata_meteo}
\title{Funkcja czyszczaca metadane}
\usage{
clean_metadata_meteo(adres, rzad = "synop", interwal = "terminowe")
}
\arguments{
\item{adres}{adres URL do pliku z metadanymi}

\item{rzad}{rzad}

\item{interwal}{interwal czasowy}
}
\description{
Tylko do wewnetrznego uzytku...
Na potrzeby funkcji 'meteo_metadane()'
}
\examples{
# przykladowe pobranie:
\dontrun{
  clean_metadata_meteo(adres = "https://dane.imgw.pl/data/dane_pomiarowo_obserwacyjne/dane_meteorologiczne/dobowe/synop/s_d_format.txt", rzad="synop", interwal="terminowe")
}

}
