% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro.R
\name{hydro}
\alias{hydro}
\title{Hydrological data}
\usage{
hydro(interval, year, value = NULL)
}
\arguments{
\item{interval}{\itemize{
\item temporal resolution of the data ("daily" , "monthly", or "semiannual_and_annual")
}}

\item{year}{vector of years (e.g., 1966:2000)}

\item{value}{type of data (can be: "State=H (default)", "Flow=Q", "Temperature=T")}
}
\description{
Pobranie danych hydrologicznych dla okresów "dobowe" , "miesieczny" , "polroczne_i_roczne"
udostepnionych w zbiorze danepubliczne.imgw.pl
}
\examples{
\dontrun{
  x <- hydro("monthly", year = 1999)
  head(x)
}
}
