% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro.R
\name{hydro}
\alias{hydro}
\title{Hydrological data}
\usage{
hydro(interval = "miesieczne", year = 1966:2000, coords = FALSE,
  value = NULL)
}
\arguments{
\item{interval}{argument accepting the value 'terminowe' , dobowe' (default), or 'miesieczne'.  Indicates for which time resolution metadata for environment R should be prepared}

\item{year}{vector of years (e.g. 1966:2000)}

\item{coords}{add coordinates for the station (logical value TRUE or FALSE)}

\item{value}{type of data ( can be: "State=H (default)" , "Flow=Q" , "Temperature=T")}
}
\description{
Pobranie danych hydrologicznych dla okresów "dobowe" , "miesieczny" , "polroczne_i_roczne"
udostepnionych w zbiorze danepubliczne.imgw.pl
}
\examples{
\dontrun{

}
}
