% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro.R
\name{hydro}
\alias{hydro}
\title{Hydrological data}
\usage{
hydro(interwal = "miesieczne", rok = 1966:2000, coords = FALSE,
  value = NULL)
}
\arguments{
\item{interwal}{interwał stacji (do wyboru: "dobowe" , "miesieczny"(domyslnie) , "polroczne_i_roczne")}

\item{rok}{wektor dla wybranych lat (np. 1966:2000)}

\item{coords}{czy dodac wspolrzedne (WGS84) i wysokosc stacji? (domyslnie coords = FALSE)}

\item{value}{rodzaj danych (do wyboru: "Stan=H (domyslnie)" , "Przepływ=Q" , "Temperatura=T")}
}
\description{
Pobranie danych hydrologicznych dla okresów "dobowe" , "miesieczny" , "polroczne_i_roczne"
udostepnionych w zbiorze danepubliczne.imgw.pl
}
\examples{
\dontrun{

}

}
