% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_annual.R
\name{hydro_annual}
\alias{hydro_annual}
\title{Semi-annual and annual hydrological data}
\usage{
hydro_annual(year, coords = FALSE, value = "H")
}
\arguments{
\item{year}{vector of years (e.g., 1966:2000)}

\item{coords}{add coordinates of the station (logical value TRUE or FALSE)}

\item{value}{type of data (can be: state - "H" (default), flow - "Q", or temperature - "T")}
}
\description{
Downloading hydrological data for the semi-annual and annual period
available in the danepubliczne.imgw.pl collection
}
\examples{
\dontrun{
  yearly <- hydro_annual(year = 2000, value = "H")
  head(yearly)
}

}
