% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydro_monthly.R
\name{hydro_monthly}
\alias{hydro_monthly}
\title{Monthly hydrological data}
\usage{
hydro_monthly(year = 1966:2000, coords = FALSE)
}
\arguments{
\item{year}{vector of years (e.g. 1966:2000)}

\item{coords}{add coordinates for the station (logical value TRUE or FALSE)}
}
\description{
Downloading hydrological data for the monthly period, available in the collection danepubliczne.imgw.pl
}
\examples{
\dontrun{
  monthly <- hydro_monthly(year = 2000)
  head(mies)
}

}
