\name{imgw-package}
\alias{imgw-package}
\alias{imgw}
\docType{package}
\title{
\packageTitle{imgw}
}
\description{
\packageDescription{imgw}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{imgw}
\packageIndices{imgw}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{imgw}

Maintainer: \packageMaintainer{imgw}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
