% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imgw-package.R
\docType{package}
\name{imgw-package}
\alias{imgw}
\alias{imgw-package}
\title{imgw: Polish Meteorological and Hydrological Data}
\description{
Download Polish meteorological and hydrological data from the Institute of Meteorology and Water Management - National Research Institute (<https://dane.imgw.pl/>).
    This package also allows for adding geographical coordinates for each observation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://imgw.ml}
  \item Report bugs at \url{https://github.com/bczernecki/imgw/issues}
}

}
\author{
\strong{Maintainer}: Bartosz Czernecki \email{nwp@amu.edu.pl} (\href{https://orcid.org/0000-0001-6496-1386}{ORCID})

Authors:
\itemize{
  \item Arkadiusz Głogowski \email{arkadiusz.glogowski@upwr.edu.pl} (\href{https://orcid.org/0000-0002-7587-8892}{ORCID})
  \item Jakub Nowosad \email{nowosad.jakub@gmail.com} (\href{https://orcid.org/0000-0002-1057-3721}{ORCID})
}

Other contributors:
\itemize{
  \item IMGW-PIB (source of the data) [contributor]
}

}
\keyword{internal}
