% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadane.R
\name{metadane}
\alias{metadane}
\title{Funkcja metadane}
\usage{
metadane(okres)
}
\arguments{
\item{okres}{-  argumenty przyjmujący wartość: 'dobowe', 'miesieczne' badz 'terminowe'.  Oznacza dla jakiej rozdzielczosci czasowej maja zostac przygotowane metadane dla srodowiska R}
}
\description{
Funkcja metadane
}
\examples{
# przykladowe uzycie:
meta <- metadane(okres="dobowe")
meta <- metadane(okres="terminowe")
meta <- metadane(okres="miesieczne")

}

