% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadane.R
\name{metadane}
\alias{metadane}
\title{Pobranie opisu (metadanych) do danych meteorologicznych  udostępnionych w repozytorium danepubliczne.imgw.pl
Domyślnie funkcja zwraca listę bądź ramkę danych dla wybranego podzbioru (w zależności od radzaj)}
\usage{
metadane(interwal, rzad)
}
\arguments{
\item{interwal}{\itemize{
\item Argument przyjmujący wartość: 'terminowe' , dobowe' (domyślnie), lub 'miesieczne'.  Oznacza dla jakiej rozdzielczosci czasowej maja zostac przygotowane metadane dla srodowiska R
}}

\item{rzad}{\itemize{
\item Argument przyjmujący wartość: 'synop' (domyślnie) lub 'klimat'.  Oznacza dla jakiego rzadu/rzędu staji mają zostać przygotowane metadane.
}}
}
\description{
Pobranie opisu (metadanych) do danych meteorologicznych  udostępnionych w repozytorium danepubliczne.imgw.pl
Domyślnie funkcja zwraca listę bądź ramkę danych dla wybranego podzbioru (w zależności od radzaj)
}
\examples{
# przykladowe uzycie:
\dontrun{
meta <- metadane(interwal = "dobowe", rzad = "synop")
meta <- metadane(interwal = "terminowe", rzad = "synop")
meta <- metadane(interwal = "miesieczne", rzad = "opad")
}
}
