% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo.R
\name{meteo}
\alias{meteo}
\title{Meteorological data}
\usage{
meteo(interval, rank, year, status = FALSE, coords = FALSE,
  col_names = "short", ...)
}
\arguments{
\item{interval}{temporal resolution of the data ("hourly", "daily", "monthly")}

\item{rank}{rank of the stations ("synop", "climate", "precip")}

\item{year}{vector of years (e.g., 1966:2000)}

\item{status}{leave the columns with measurement and observation statuses (default status = FALSE - i.e. the status columns are deleted)}

\item{coords}{add coordinates of the station (logical value TRUE or FALSE)}

\item{col_names}{three types of column names possible: "short" - default, values with shorten names, "full" - full English description, "polish" - original names in the dataset}

\item{...}{other parameters that may be passed to 'shortening' function that shortens column names}
}
\description{
Downloading hourly, daily, and monthly meteorological data from the SYNOP / CLIMATE / PRECIP stations available in the danepubliczne.imgw.pl collection
}
\examples{
\dontrun{
  x <- meteo("monthly", rank = "synop", year = 2018, coords = TRUE)
  head(x)
}
}
