% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo_daily.R
\name{meteo_daily}
\alias{meteo_daily}
\title{Daily meteorological data}
\usage{
meteo_daily(rank, year, status = FALSE, coords = FALSE)
}
\arguments{
\item{rank}{rank of the stations ("synop", "climate", or "precip")}

\item{year}{vector of years (e.g., 1966:2000)}

\item{status}{leave the columns with measurement and observation statuses (default status = FALSE - i.e. the status columns are deleted)}

\item{coords}{add coordinates of the station (logical value TRUE or FALSE)}
}
\description{
Downloading daily (meteorological) data from the SYNOP / CLIMATE / PRECIP stations available in the danepubliczne.imgw.pl collection
}
\examples{
\dontrun{
  daily <- meteo_daily(rank = "climate", year = 2000)
  head(daily)
}

}
