% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo_hourly.R
\name{meteo_hourly}
\alias{meteo_hourly}
\title{Hourly meteorological data}
\usage{
meteo_hourly(rank, year, status = FALSE, coords = FALSE,
  short = TRUE, ...)
}
\arguments{
\item{rank}{rank of the stations ("synop", "climate", or "precip")}

\item{year}{vector of years (e.g., 1966:2000)}

\item{status}{leave the columns with measurement and observation statuses (default status = FALSE - i.e. the status columns are deleted)}

\item{coords}{add coordinates of the station (logical value TRUE or FALSE)}

\item{short}{shortening column names (logical value TRUE or FALSE)}

\item{...}{other parameters that may be passed to 'shortening' function that shortens column names}
}
\description{
Downloading hourly (meteorological) data from the SYNOP / CLIMATE / PRECIP stations available in the danepubliczne.imgw.pl collection
}
\examples{
\dontrun{
  hourly <- meteo_hourly(rank = "climate", year = 1984)
  head(hourly)
}

}
