% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo_metadata.R
\name{meteo_metadata}
\alias{meteo_metadata}
\title{Meteorological metadata}
\usage{
meteo_metadata(interval, rank)
}
\arguments{
\item{interval}{\itemize{
\item Argument accepting the value 'hourly', 'daily' (default), or 'monthly'.  Indicates for which time resolution metadata for environment R should be prepared
}}

\item{rank}{rank of station ("synop" , "climate" , "precip")}
}
\description{
Downloading the description (metadata) to the meteorological data available in the danepubliczne repository.imgw.pl
By default, the function returns a list or data frame for a selected subset (depending on how to cop#`
}
\examples{
\dontrun{
  meta <- meteo_metadata(interval = "hourly", rank = "synop")
  meta <- meteo_metadata(interval = "daily", rank = "synop")
  meta <- meteo_metadata(interval = "monthly", rank = "precip")
}
}
