% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo_miesieczne.R
\name{meteo_miesieczne}
\alias{meteo_miesieczne}
\title{Pobranie danych miesiecznych (meteorologicznych) ze stacji SYNOP/KLIMAT/OPAD udostepnionych w zbiorze danepubliczne.imgw.pl}
\usage{
meteo_miesieczne(rzad = "synop")
}
\arguments{
\item{rzad}{rzad}
}
\description{
Pobranie danych miesiecznych (meteorologicznych) ze stacji SYNOP/KLIMAT/OPAD udostepnionych w zbiorze danepubliczne.imgw.pl
}
\examples{
\dontrun{
  mies <- meteo_miesieczne(rzad = "klimat")
  head(mies)
}

}
