% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo_miesieczne.R
\name{meteo_miesieczne}
\alias{meteo_miesieczne}
\title{Funkcja pobierająca dane miesięczne ze stacji SYNOP/KLIMAT dostępne w zbiorze danepubliczne.imgw.pl}
\usage{
meteo_miesieczne(okres = "miesieczne", rodzaj = "synop")
}
\description{
Funkcja pobierająca dane miesięczne ze stacji SYNOP/KLIMAT dostępne w zbiorze danepubliczne.imgw.pl
}
\examples{

}
