% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo_monthly.R
\name{meteo_monthly}
\alias{meteo_monthly}
\title{Monthly meteorological data}
\usage{
meteo_monthly(rank, year, status = FALSE, coords = FALSE)
}
\arguments{
\item{rank}{rank of the stations ("synop", "klimat", or "opad")}

\item{year}{vector of years (e.g., 1966:2000)}

\item{status}{leave the columns with measurement and observation statuses (default status = FALSE - i.e. the status columns are deleted)}

\item{coords}{add coordinates of the station (logical value TRUE or FALSE)}
}
\description{
Downloading monthly (meteorological) data from the SYNOP / CLIMATE / PRECIP stations available in the danepubliczne.imgw.pl collection
}
\examples{
\dontrun{
  monthly <- meteo_monthly(rank = "climate", year = 1969)
  head(monthly)
}

}
