% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pobierz.R
\name{pobierz}
\alias{pobierz}
\title{Funkcja pobierająca dane z serwisu dane.imgw.pl}
\usage{
pobierz(data_start, data_end, user_pass, stacja, kod)
}
\arguments{
\item{data_start}{data początkowa od której beda pobierane dane. Obiekt klasy Date}

\item{data_end}{data końcowa do ktorej beda pobierane. Obiekt klasy Date. Domyslnie funkcja rozszerzy okres pobierania o tydzien w przod i w tyl}

\item{user_pass}{login i token ze strony https://dane.imgw.pl. Zmienna w postaci email:token, np. bczernecki@gmail.com:68a60056}

\item{stacja}{kod stacji wg strony https://dane.imgw.pl}

\item{kod}{kod pobieranego elementu meteorologicznego/hydrologicznego}
}
\description{
Pozwala na zautomatyzowanie procesu pobierania danych meteorologicznych i hydrologicznych z oficjalnego repozytorium Instytutu Meteorologii i Gospodarki Wodnej - Państwowego Instytutu Badawczego (https://dane.imgw.pl)
}
\examples{
# wymagane ustalenie ponizszych parametrow pobierania:

user_pass <- "" 
data_start <- "2016-12-26" 
data_end <- "2017-02-20" 
stacja <- "352160330" # przykladowe id dla Poznania; numer mozna pobrac ze strony https://dane.imgw.pl
kod <- "B100B008CD" # srednia dobowa temp. powietrza synop; j.w.
pobierz(data_start,data_end,user_pass,stacja,kod)
}

